backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
httpx==0.25.1
beautifulsoup4==4.12.2
feedparser==6.0.10
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiohttp==3.8.6
aiocache==0.12.2
nltk==3.8.1
scikit-learn==1.3.2
pytest==7.4.3 
email-validator==1.5.1